requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-rest-framework==0.1.0
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
